fix(PL.Validation): Solve the cancelation of form closing event when invalid inputs exist
Problem:
The previous approach was relying on "this.ValidateChildren" which needs "e.Cancel=true" in Validating events,
to return correct boolean, and it worked correctly, until the following scenario happened:
When the user clicks the X button, the form calls Validating events, and if there are invalid inputs, it cancels the
closing event of the form, keeping the user "trapped" in this form and unable to close it unless all the inputs
are valid, which is irrational.

Solution:
Deleted "e.Cancel=true" from Validating events, so closing event won't be canceled,
but "this.ValidateChildren" won't return correct result any more, so i implemented
a new extension method "HasErrors" for ErrorProvider control, to make decision upon its output.

Benefits:
- user able to close the form any time.
- immediate error feedback when user changes focus between controls.
- Input validation still functioning on submission.